refactoring for best practice
